pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["edoch <edo_chio@hotmail.it>"]
classifiers = ["Private :: Do not Upload"]
description = "Project for the Machine Learning and Deep Learning course of Data Science and Engineering Master Degree at Politecnico di Torino"
exclude = ["**"]
license = "Proprietary"
name = "waste-semantic-segmentation"
readme = "README.md"
version = "0.0.0"

[tool.poetry.dependencies]
hydra-core = "^1.3.2"
python = ">=3.10,<3.13"
torch = "^2.1.2"
torchvision = "^0.16.2"
tqdm = "^4.66.1"

[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"
ipykernel = "^6.26.0"
pre-commit = "^3.5.0"
ruff = "^0.1.3"

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I"
]
